# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.13.0 🍺
 * Do not edit manually.
 * Api
 * Lectio — Foreign Language Reader API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface VisitInput {
  /**
     * Stable per-browser identifier (random, stored in localStorage)
     * @minLength 1
     * @maxLength 128
     */
  visitorId: string;
  /**
     * The path that was visited
     * @minLength 1
     * @maxLength 512
     */
  path: string;
  /**
     * document.referrer, if any
     * @maxLength 1024
     * @nullable
     */
  referrer?: string | null;
}

export interface VisitAck {
  recorded: boolean;
}

export interface VisitStats {
  /** All-time visit count */
  total: number;
  /** Distinct visitor count */
  uniqueVisitors: number;
  /** Visits in the last 24 hours */
  last24h: number;
  /** Visits in the last 7 days */
  last7d: number;
}

export interface TextSearchRequest {
  /** e.g. 'The Prince by Machiavelli, chapter 1' */
  query: string;
}

export interface TranslationOptions {
  /** Target gloss/translation language (defaults to the text's target language / English) */
  language?: string;
}

export interface Text {
  id: number;
  title: string;
  author: string;
  /** e.g. 'Italian', 'Latin', 'French' */
  language: string;
  /**
     * The language to translate to, e.g. 'English'
     * @nullable
     */
  targetLanguage?: string | null;
  /**
     * URL of the open-source text
     * @nullable
     */
  sourceUrl: string | null;
  /** @nullable */
  description?: string | null;
  paragraphCount: number;
  /**
     * Approximate publication year (negative for BCE, e.g. -750 for 750 BCE)
     * @nullable
     */
  publicationYear?: number | null;
  /**
     * Commonly used English title (e.g. 'Crime and Punishment' for 'Преступление и наказание')
     * @nullable
     */
  englishTitle?: string | null;
  /**
     * Author's name in Latin script as commonly used in English (e.g. 'Fyodor Dostoevsky')
     * @nullable
     */
  englishAuthor?: string | null;
  /**
     * Stable catalog identifier for seeded texts (e.g. 'Aeneis'), used for deep links
     * @nullable
     */
  catalogKey?: string | null;
  /**
     * For English texts: 'British' or 'American', used to split the library
     * @nullable
     */
  nationality?: string | null;
  createdAt: string;
}

export interface TextWithProgress {
  id: number;
  title: string;
  author: string;
  language: string;
  paragraphCount: number;
  completedCount: number;
  /** @nullable */
  lastParagraphIndex?: number | null;
  /** @nullable */
  lastAccessedAt: string | null;
}

export interface TextStats {
  textId: number;
  totalParagraphs: number;
  completedParagraphs: number;
  percentComplete: number;
}

export interface ParagraphSummary {
  id: number;
  textId: number;
  index: number;
  originalText: string;
  completed: boolean;
}

export interface Paragraph {
  id: number;
  textId: number;
  index: number;
  originalText: string;
  completed: boolean;
  /**
     * Cached interlinear translation
     * @nullable
     */
  interlinearTranslation?: string | null;
  /**
     * Cached full literary translation
     * @nullable
     */
  fullTranslation?: string | null;
}

export interface VocabularyEntry {
  original: string;
  translation: string;
  count: number;
  firstParagraphIndex: number;
}

export interface VocabularyList {
  textId: number;
  /** Highest paragraph index included in the aggregation (-1 if none) */
  throughParagraphIndex: number;
  entries: VocabularyEntry[];
}

export interface Flashcard {
  /** The word as it appears in the text */
  word: string;
  /** Short English definition (headword meaning) */
  definition: string;
  /** A single emoji memory hook */
  icon: string;
  /** Grammatical form of the word (may be empty) */
  inflection: string;
  count: number;
  firstParagraphIndex: number;
}

export interface FlashcardDeck {
  textId: number;
  /** Highest paragraph index included (-1 if none) */
  throughParagraphIndex: number;
  cards: Flashcard[];
}

export interface PublicQuizQuestion {
  id: number;
  /** translation | vocab | grammar */
  kind: string;
  prompt: string;
  options: string[];
}

export interface PublicQuiz {
  textId: number;
  paragraphIndex: number;
  paragraphText: string;
  questions: PublicQuizQuestion[];
}

export interface QuizAnswer {
  id: number;
  chosenIndex: number;
}

export interface QuizSubmission {
  answers: QuizAnswer[];
}

export interface QuizResultItem {
  id: number;
  kind: string;
  prompt: string;
  options: string[];
  correctIndex: number;
  chosenIndex: number;
  correct: boolean;
  explanation: string;
}

export interface QuizResult {
  textId: number;
  paragraphIndex: number;
  score: number;
  total: number;
  items: QuizResultItem[];
}

export interface ReviewWeakItem {
  kind: string;
  prompt: string;
  explanation?: string;
  correctAnswer?: string;
  missedCount: number;
  lastSeenAt: string;
  textId?: number;
  paragraphIndex?: number;
  textTitle?: string;
}

export interface ReviewSummary {
  totalAttempts: number;
  totalScore: number;
  totalPossible: number;
  weakItems: ReviewWeakItem[];
}

export type WordLookupAnalysisFeaturesItem = {
  label: string;
  value: string;
};

export interface WordLookupAnalysis {
  lemma?: string;
  features: WordLookupAnalysisFeaturesItem[];
}

export interface WordLookup {
  word: string;
  language: string;
  /** perseus | ai | wiktionary | none */
  source: string;
  sourceUrl: string;
  analyses: WordLookupAnalysis[];
}

export interface InterlinearWord {
  original: string;
  translation: string;
  /** Romanized pronunciation for non-Latin scripts (Greek, Cyrillic, etc.). Omitted for Latin-script languages. */
  transliteration?: string;
}

export interface InterlinearTranslation {
  paragraphId: number;
  originalText: string;
  words: InterlinearWord[];
}

export interface FullTranslation {
  paragraphId: number;
  originalText: string;
  translatedText: string;
}

export interface Scansion {
  paragraphId: number;
  originalText: string;
  scannedText: string;
}

export interface ParagraphAudio {
  paragraphId: number;
  format: string;
  audioBase64: string;
}

export interface Progress {
  id: number;
  textId: number;
  paragraphIndex: number;
  completed: boolean;
  updatedAt: string;
}

export interface ProgressInput {
  textId: number;
  paragraphIndex: number;
  completed: boolean;
}

export interface SubscriptionStatus {
  active: boolean;
  status?: string | null;
}

export interface CheckoutUrl {
  url: string;
}

export interface DonationInput {
  /**
     * Donation amount in US cents (min $1, max $10,000).
     * @minimum 100
     * @maximum 1000000
     */
  amountCents: number;
}

export type LookupWordParams = {
/**
 * Language label as stored on the text (e.g. "Latin", "Ancient Greek")
 */
lang: string;
word: string;
};