# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Lectio — Foreign Language Reader API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /texts/{textId}/paragraphs:
    get:
      operationId: listParagraphs
      tags: [paragraphs]
      summary: List all paragraphs for a text with completion status
      parameters:
        - name: textId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of paragraphs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ParagraphSummary"
  /texts/{textId}/paragraphs/{index}:
    get:
      operationId: getParagraph
      tags: [paragraphs]
      summary: Get a specific paragraph (original text only)
      parameters:
        - name: textId
          in: path
          required: true
          schema:
            type: integer
        - name: index
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Paragraph details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Paragraph"
        "404":
          description: Not found
...
  /texts/{textId}/paragraphs/{index}/interlinear:
    post:
      operationId: getInterlinearTranslation
      tags: [paragraphs]
      summary: Generate or retrieve the interlinear (word-for-word) translation
      parameters:
        - name: textId
          in: path
          required: true
          schema:
            type: integer
        - name: index
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TranslationOptions"
      responses:
        "200":
          description: Interlinear translation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InterlinearTranslation"
        "404":
          description: Not found
  /texts/{textId}/paragraphs/{index}/translation:
    post:
      operationId: getFullTranslation
      tags: [paragraphs]
      summary: Generate or retrieve the full literary translation
      parameters:
        - name: textId
          in: path
          required: true
          schema:
            type: integer
        - name: index
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TranslationOptions"
      responses:
        "200":
          description: Full literary translation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FullTranslation"
        "404":
          description: Not found
...
  /texts/{textId}/paragraphs/{index}/scansion:
    post:
      operationId: getScansion
      tags: [paragraphs]
      summary: Generate or retrieve scansion marks (macrons/breves) for a verse paragraph
      parameters:
        - name: textId
          in: path
          required: true
          schema:
            type: integer
        - name: index
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Scansion-marked text
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scansion"
        "404":
          description: Not found
  /texts/{textId}/paragraphs/{index}/audio:
    post:
      operationId: getParagraphAudio
      tags: [paragraphs]
      summary: Generate or retrieve spoken audio of a paragraph in its original language
      parameters:
        - name: textId
          in: path
          required: true
          schema:
            type: integer
        - name: index
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Spoken audio of the paragraph
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParagraphAudio"
        "404":
          description: Not found
...
  /texts/{textId}/flashcards:
    get:
      operationId: getTextFlashcards
      tags: [texts]
      summary: Vocabulary flashcards for words seen so far in this text
      parameters:
        - name: textId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Flashcard deck
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlashcardDeck"
...
  /donate/checkout:
    post:
      operationId: createDonationCheckout
      tags: [donate]
      summary: Create a Stripe Checkout session for a one-time donation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DonationInput"
      responses:
        "200":
          description: Checkout URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutUrl"
        "400":
          description: Invalid donation amount
...
components:
  schemas:
...
    TranslationOptions:
      type: object
      properties:
        language:
          type: string
          description: "Target gloss/translation language (defaults to the text's target language / English)"
...
    ParagraphSummary:
      type: object
      required: [id, textId, index, originalText, completed]
      properties:
        id:
          type: integer
        textId:
          type: integer
        index:
          type: integer
        originalText:
          type: string
        completed:
          type: boolean
    Paragraph:
      type: object
      required: [id, textId, index, originalText, completed]
      properties:
        id:
          type: integer
        textId:
          type: integer
        index:
          type: integer
        originalText:
          type: string
        completed:
          type: boolean
        interlinearTranslation:
          type: ["string", "null"]
          description: "Cached interlinear translation"
        fullTranslation:
          type: ["string", "null"]
          description: "Cached full literary translation"
...
    Flashcard:
      type: object
      required: [word, definition, icon, inflection, count, firstParagraphIndex]
      properties:
        word:
          type: string
          description: The word as it appears in the text
        definition:
          type: string
          description: Short English definition (headword meaning)
        icon:
          type: string
          description: A single emoji memory hook
        inflection:
          type: string
          description: Grammatical form of the word (may be empty)
        count:
          type: integer
        firstParagraphIndex:
          type: integer
    FlashcardDeck:
      type: object
      required: [textId, throughParagraphIndex, cards]
      properties:
        textId:
          type: integer
        throughParagraphIndex:
          type: integer
          description: Highest paragraph index included (-1 if none)
        cards:
          type: array
          items:
            $ref: "#/components/schemas/Flashcard"
...
    InterlinearWord:
      type: object
      required: [original, translation]
      properties:
        original:
          type: string
        translation:
          type: string
        transliteration:
          type: string
          description: Romanized pronunciation for non-Latin scripts (Greek, Cyrillic, etc.). Omitted for Latin-script languages.
    InterlinearTranslation:
      type: object
      required: [paragraphId, words, originalText]
      properties:
        paragraphId:
          type: integer
        originalText:
          type: string
        words:
          type: array
          items:
            $ref: "#/components/schemas/InterlinearWord"
    FullTranslation:
      type: object
      required: [paragraphId, originalText, translatedText]
      properties:
        paragraphId:
          type: integer
        originalText:
          type: string
        translatedText:
          type: string
    Scansion:
      type: object
      required: [paragraphId, originalText, scannedText]
      properties:
        paragraphId:
          type: integer
        originalText:
          type: string
        scannedText:
          type: string
    ParagraphAudio:
      type: object
      required: [paragraphId, format, audioBase64]
      properties:
        paragraphId:
          type: integer
        format:
          type: string
        audioBase64:
          type: string
...
    CheckoutUrl:
      type: object
      required: [url]
      properties:
        url:
          type: string
    DonationInput:
      type: object
      required: [amountCents]
      properties:
        amountCents:
          type: integer
          minimum: 100
          maximum: 1000000
          description: "Donation amount in US cents (min $1, max $10,000)."