# lib/api-zod/src/generated/types/donationInput.ts
/**
 * Generated by orval v8.13.0 🍺
 * Do not edit manually.
 * Api
 * Lectio — Foreign Language Reader API
 * OpenAPI spec version: 0.1.0
 */

export interface DonationInput {
  /**
     * Donation amount in US cents (min $1, max $10,000).
     * @minimum 100
     * @maximum 1000000
     */
  amountCents: number;
}