# lib/api-zod/src/generated/types/paragraphSummary.ts
/**
 * Generated by orval v8.13.0 🍺
 * Do not edit manually.
 * Api
 * Lectio — Foreign Language Reader API
 * OpenAPI spec version: 0.1.0
 */

export interface ParagraphSummary {
  id: number;
  textId: number;
  index: number;
  originalText: string;
  completed: boolean;
}